worker/requirements.txt
flask
flask-cors
opencv-python
numpy
requests
python-dotenv
minio
ffmpeg-python
ultralytics